requirements/base.txt
# Base requirements to run the project

# Django
Django==4.2.2

# Database
psycopg2-binary==2.9.6

# Pagination
django-el-pagination==4.0.0

# PWA
django-pwa==1.1.0

# TASKS
celery==5.3.1
redis==4.6.0

# ENVIRONMENT
python-dotenv==1.0.0

# EMAILS
sendgrid==6.10.0

# DEPLOYMENT
gunicorn==21.2.0

# DJANGO DEVELOPMENT UTILITIES
django-extensions==3.2.3

# STATIC FILES
whitenoise==6.5.0

# TRANSLATIONS
django-modeltranslation==0.18.11

# FORMS
django-select2==8.1.2

# CACHES
django-redis==5.3.0

# WEB PUSH
pywebpush==1.14.0
